move asyncCanClockOut out of Main VM
