+++ app/src/main/java/com/rockthevote/grommet/ui/eventFlow/SessionTimeTrackingViewModel.kt
@@ -3,20 +3,23 @@ package com.rockthevote.grommet.ui.eventFlow
 import androidx.lifecycle.*
 import com.hadilq.liveevent.LiveEvent
 import com.rockthevote.grommet.data.db.dao.PartnerInfoDao
+import com.rockthevote.grommet.data.db.dao.RegistrationDao
 import com.rockthevote.grommet.data.db.dao.SessionDao
 import com.rockthevote.grommet.util.coroutines.DispatcherProvider
 import com.rockthevote.grommet.util.coroutines.DispatcherProviderImpl
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 import timber.log.Timber
 
 /**
  * Created by Mechanical Man on 5/30/20.
  */
 class SessionTimeTrackingViewModel(
-        private val dispatchers: DispatcherProvider = DispatcherProviderImpl(),
-        private val partnerInfoDao: PartnerInfoDao,
-        private val sessionDao: SessionDao
+    private val dispatchers: DispatcherProvider = DispatcherProviderImpl(),
+    private val partnerInfoDao: PartnerInfoDao,
+    private val sessionDao: SessionDao,
+    private val registrationDao: RegistrationDao
 ) : ViewModel() {
 
     private val coroutineExceptionHandler = CoroutineExceptionHandler { _, throwable ->
@@ -53,6 +56,26 @@ class SessionTimeTrackingViewModel(
         }
     }
 
+    /**
+     * Asynchronously determines if the user can clock out. Calls [successCallback]
+     * when the user can logout, [failCallback] when the user cannot.
+     *
+     * Note: Coroutines cannot be used from Java, so callbacks are necessary
+     */
+    fun asyncCanClockOut(successCallback: () -> Unit, failCallback: () -> Unit) {
+        viewModelScope.launch(dispatchers.io) {
+            val canClockOut = registrationDao.getAll().isEmpty()
+
+            withContext(dispatchers.main) {
+                if (canClockOut) {
+                    successCallback()
+                } else {
+                    failCallback()
+                }
+            }
+        }
+    }
+
     fun clearSession() {
         viewModelScope.launch(dispatchers.io + coroutineExceptionHandler) {
             sessionDao.clearAllSessionInfo()
@@ -71,12 +94,13 @@ class SessionTimeTrackingViewModel(
 
 class SessionTimeTrackingViewModelFactory(
         private val partnerInfoDao: PartnerInfoDao,
-        private val sessionDao: SessionDao
+        private val sessionDao: SessionDao,
+        private val registrationDao: RegistrationDao
 ) : ViewModelProvider.Factory {
     override fun <T : ViewModel?> create(modelClass: Class<T>): T {
         val dispatchers = DispatcherProviderImpl()
 
         @Suppress("UNCHECKED_CAST")
-        return SessionTimeTrackingViewModel(dispatchers, partnerInfoDao, sessionDao) as T
+        return SessionTimeTrackingViewModel(dispatchers, partnerInfoDao, sessionDao, registrationDao) as T
     }
 }
